+++ typings/index.d.ts
@@ -1,0 +1,2 @@
+/// <reference path="globals/glob/index.d.ts" />
+/// <reference path="globals/minimatch/index.d.ts" />