+++ CHANGELOG.md
@@ -1,2 +1,10 @@
+## Version 0.0.3
+* add command `tortoise-svn...` which can select TortoiseSVN action and target dir or file
+* add menu `tortoise-svn...` to `explorer/context` and `editor/context` 
+* add `keybindings` `alt+s m`: tortoise-svn...
+
+## Version 0.0.2
+* add more menu
+
 ## Version 0.0.1
 * Initial release